PtFEM.jl Documentation

Programs

PtFEM.p41 — Method.

Method p41

One dimensional analysis of axially loaded elastic rods using 2-node rod elements.

Constructors

p41(data::Dict)
p41(m::jFEM, data::Dict) # Used to re-use factored global stiffness matrix

Arguments

* `m`    : Previously created jFEM model
* `data` : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::0.0:0.1:1.0                                : x-coordinate vector

Optional additional data dictionary keys

* penalty = 1e20               : Penalty used for fixed degrees of freedoms
* etype::Vector{Int64}         : Element material vector if np_types > 1
* eq_nodal_forces_and_moments  : Contribution of distributed loads to loaded_nodes

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement             : List of available structural element types
?Rod                           : Help on a Rod structural element
?FiniteElement                 : List finite element types
?Line                          : Help on Line finite element

source

PtFEM.p42 — Method.

Method p42

Analysis of elastic pin-jointed frames using 2-node rod elements in 2- or 3-dimensions

Constructors

p42(data::Dict)

Arguments

* `data` : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of structural element
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::Vector{Float64}                            : x coordinate vector
* y_coords::Vector{Float64}                            : y coordinate vector
* g_num::Array{Int64,2}                                : Element node connections

Optional dictionary keys

* penalty::Float64             : Penalty for fixed freedoms
* etype::Vector{Int64}         : Element material vector
* z_coords::Vector{Float64}    : z coordinate vector (3D)
* eq_nodal_forces_and_moments  : Contribution of distributed loads to loaded_nodes

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement  : List structural element types
?Frame              : Help on a Rod structural fin_el
?FiniteElement      : List finite element types
?Line               : Help on Line finite element

source

PtFEM.p43 — Method.

p43

Method for static equilibrium analysis of a beam.

Constructors

p43(data::Dict)

Arguments

* `data` : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::LinSpace{Float64}                          : x coordinate vector
* g_num::Array{Int64,2}                                : Element node connections
* fixed_freedoms::Array{Tuple{Vector{Int64}}           : Fixed freedoms

Optional dictionary keys

* etype::Vector{Int64}                                 : Element material vector
* penalty::Float64                                     : Penalty for fixed freedoms
* eq_nodal_forces_and_moments                          : Equivalent nodal loads

Examples

using PtFEM

data = Dict(
  # Beam(ndim, nst, nxe, nip, direction, finite_element(nod, nodof), axisymmet
  :struc_el => Beam(2, 1, 4, 1, :x, Line(2, 2), false),
  :properties => [4.0e4; 2.0e4],
  :etype => [1, 1, 2, 2],
  :x_coords => [0.0, 2.5, 5.0, 8.0, 10.0],
  :support => [
      (1, [0 1]),
      (4, [0 1])
    ],
  :loaded_nodes => [
      (2, [-20.0 0.0]),
      (3, [-6.0 -3.0]),
      (4, [-8.8 2.2]),
      (5, [-1.2 0.5333])
    ],
  :fixed_freedoms => [
      (1, 2, -0.001),
      (3, 1, -0.005)
    ],
  :penalty => 1e19,
  :eq_nodal_forces_and_moments => [
    (3, [-6.0 -3.0 -6.0 3.0]),
    (4, [-2.8 -0.8 -1.2  0.5333])
  ]
)

fem, dis_dt, fm_dt = p43(data)

println("Displacements:")
dis_dt |> display
println()

println("Actions:")
fm_dt |> display
println()

Related help

?StructuralElement  : Help on structural elements
?Rod                : Help on a Rod structural fin_el
?FiniteElement      : Help on finite element types

source

Structural Elements

PtFEM.StructuralElement — Type.

StructuralElement

Abstract structural element type.

Type

abstract StructuralElement

Subtypes

* Rod::StructuralElement          : Rod(nxe, np_types, nip, fin_el)
* Beam::StructuralElement         : Beam(nod, nodof)
* Frame::StructuralElement        : Frame(nod, nodof)
* Plane::StructuralElement        : Plane(nod, nodof)
* Solid::StructuralElement        : Solid(nod, nodof)
* GenericSolid::StructuralElement : GenericSolid(nod, nodof)

Related help

?FiniteElement                   : Show all finite elements
?Rod                              : Help on Rod structural element
?Beam                             : Help on Beam structural element
?Frame                            : Help on Frame structural element
?Plane                            : Help on Plane structural element
?Solid                            : Help on Solid structural element
?GenericSolid                     : Help on GenericSolid structural element

source

PtFEM.Rod — Type.

Rod

Concrete 1D structural element with only axial stresses.

Constructor

Rod(nels, np_types, nip, fin_el)

Arguments

* nels::Int64             : Number of fin_els (stored in field nxe)
* np_types::Int64         : Number of different property types
* nip::Int64              : Number of integration points
* fin_el::FiniteElement   : Line(nod, nodof)

Related help

?StructuralElement  : Help on structural elements
?FiniteElement      : Help on finite element types
?Line               : Help on a Line finite element

source

PtFEM.Beam — Type.

Beam

Concrete structural element with transverse and moment loading.

Constructor

Beam(ndim, nip, fin_el)

Arguments

* ndim::Int64             : Number of dimensions
* nst::Int64              : Number of stress terms
* nxe::Int64              : Number of different property types
* nip::Int64              : Number of integration points
* direction::Symbol       : Number of integration points
* fin_el::FiniteElement   : Line(nod, nodof)
* axisymmetric::Bool      : Axisymmetric if true

Related help

?StructuralElement  : Help on structural elements
?FiniteElement      : Help on finite element types
?Line               : Help on a Line finite element

source

Finite Elements

PtFEM.FiniteElement — Type.

FiniteElement

Abstract finite element type.

Type

abstract FiniteElement

Subtypes

* Line::FiniteElement          : 1D Line(nod, nodof)
* Triangle::FiniteElement      : 2D Triangle(nod, nodof)
* Quadrilateral::FiniteElement : 2D Quadrilateral(nod, nodof)
* Hexahedron::FiniteElement    : 3D Hexahedron(nod, nodof)
* Tetrahedron::FiniteElement   : 3D Tetrahedron(nod, nodof)

source

PtFEM.Line — Type.

Line (Interval)

1D type finite element

Constructor

Line(nod, nodof)
Line(nodof)

Arguments

* nod::Int64       : Number of nodes for finite element, defaults to 2
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Triangle — Type.

Triangle

2D type finite element

Constructor

Triangle(nod, nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (3, 6, 10, 15)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM

PtFEM.fromSkyline — Method.

fromSkyline

Helper function to convert a Skyline vector to a full matrix.

Type

fromSkyline(skyline::Vector{Float64}, kdiag::Vector{Int64})

Arguments

* skyline::Vector{Float64}     : 1D Line(nod, nodof)
* kdiag::Vector{Int64}         : 2D Triangle(nod, nodof)

source

PtFEM.sparin! — Method.

Function sparin! performs Cholesky factorisation on a symmetric skyline global matrix. The vector kv is updated.

Arguments to sparin!(kv, kdiag):

kv::Vector{Float64} : Skyline vector of global stiffness matrix

kdiag::Vector{Int64} : Diagonal fin_el vector

source

Index

PtFEM.Beam

PtFEM.FiniteElement

PtFEM.Line

PtFEM.Rod

PtFEM.StructuralElement

PtFEM.Triangle

PtFEM.fromSkyline

PtFEM.p41

PtFEM.p42

PtFEM.p43

PtFEM.sparin!