PtFEM.jl Documentation

Changes with respect to the PtFEM book

Julia's convention for functions that update arguments

Custom array indices

Replacing skyline storage by Julia sparse matrices

Separate equivalent loads in data dictionary

Graphics

Initial introduction of parallel programming in Julia

Integration - for now using PtFEM's approach

Gradient descent - for now using PtFEM's approach

Derivatives - for now using PtFEM's approach

Getting started

Version approach and history

Approach

Version history

The future of the GitHub PtFEM organisation

References

TODO list

Planned work

Intended extensions but not yest planned

PtFEM.jl Documentation

Programs

Structural Elements

Finite Elements

PtFEM

Index

Introduction

Programs

PtFEM.p41 — Method.

p41

Method for static equilibrium analysis of a rod.

Constructors

p41(data::Dict)
p41(m::jFEM, data::Dict) # Used to re-use factored global stiffness matrix

Arguments

* `m`    : Previously created jFEM model
* `data` : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::LinSpace{Float64}                          : Xcoordinate vector

Optional dictionary keys

* etype::Vector{Int64}                                 : Element material vector

Examples

using PtFEM

data = Dict(
  # Rod(nels, np_types, nip, finite_element(nod, nodof))
  :struc_el => Rod(4, 1, 1, Line(2, 1)),
  :properties => [1.0e5;],
  :x_coords => linspace(0, 1, 5),
  :support => [(1, [0])],
  :loaded_nodes => [(1,[-0.625]),(2,[-1.25]),(3,[-1.25]),(4,[-1.25]),(5,[-0.625])]
)

fem, dis_dt, fm_dt = p41(data)

println("Displacements:")
dis_dt |> display
println()

println("Actions:")
fm_dt |> display
println()

Related help

?StructuralElement  : Help on structural elements
?Rod                : Help on a Rod structural fin_el
?FiniteElement      : Help on finite element types

source

Base.:* — Method.

*(x, y...)

Multiplication operator. x*y*z*... calls this function with all arguments, i.e. *(x, y, z, ...).

source

Structural Elements

PtFEM.StructuralElement — Type.

StructuralElement

Abstract structural element type.

Type

abstract StructuralElement

Subtypes

* Rod::StructuralElement          : Rod(nxe, np_types, nip, fin_el)
* Beam::StructuralElement         : Beam(nod, nodof)
* Frame::StructuralElement        : Frame(nod, nodof)
* Plane::StructuralElement        : Plane(nod, nodof)
* Solid::StructuralElement        : Solid(nod, nodof)
* GenericSolid::StructuralElement : GenericSolid(nod, nodof)

Related help

?FiniteElement                   : Show all finite elements
?Rod                              : Help on Rod structural element
?Beam                             : Help on Beam structural element
?Frame                            : Help on Frame structural element
?Plane                            : Help on Plane structural element
?Solid                            : Help on Solid structural element
?GenericSolid                     : Help on GenericSolid structural element

source

PtFEM.Rod — Type.

Rod

Concrete 1D structural element with only axial stresses.

Constructor

Rod(nels, np_types, nip, fin_el)

Arguments

* nels::Int64             : Number of fin_els (stored in field nxe)
* np_types::Int64         : Number of different property types
* nip::Int64              : Number of integration points
* fin_el::FiniteElement   : Line(nod, nodof)

Related help

?StructuralElement  : Help on structural elements
?FiniteElement      : Help on finite element types
?Line               : Help on a Line finite element

source

PtFEM.Beam — Type.

Beam

Concrete structural element with transverse and moment loading.

Constructor

Beam(ndim, nip, fin_el)

Arguments

* ndim::Int64             : Number of dimensions
* nst::Int64              : Number of stress terms
* nxe::Int64              : Number of different property types
* nip::Int64              : Number of integration points
* direction::Symbol       : Number of integration points
* fin_el::FiniteElement   : Line(nod, nodof)
* axisymmetric::Bool      : Axisymmetric if true

Related help

?StructuralElement  : Help on structural elements
?FiniteElement      : Help on finite element types
?Line               : Help on a Line finite element

source

Finite Elements

PtFEM.FiniteElement — Type.

FiniteElement

Abstract finite element type.

Type

abstract FiniteElement

Subtypes

* Line::FiniteElement          : 1D Line(nod, nodof)
* Triangle::FiniteElement      : 2D Triangle(nod, nodof)
* Quadrilateral::FiniteElement : 2D Quadrilateral(nod, nodof)
* Hexahedron::FiniteElement    : 3D Hexahedron(nod, nodof)
* Tetrahedron::FiniteElement   : 3D Tetrahedron(nod, nodof)

source

PtFEM.Line — Type.

Line (Interval)

1D type finite element

Constructor

Line(nod, nodof)
Line(nodof)

Arguments

* nod::Int64       : Number of nodes for finite element, defaults to 2
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Triangle — Type.

Triangle

2D type finite element

Constructor

Triangle(nod, nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (3, 6, 10, 15)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM

PtFEM.fromSkyline — Method.

fromSkyline

Helper function to convert a Skyline vector to a full matrix.

Type

fromSkyline(skyline::Vector{Float64}, kdiag::Vector{Int64})

Arguments

* skyline::Vector{Float64}     : 1D Line(nod, nodof)
* kdiag::Vector{Int64}         : 2D Triangle(nod, nodof)

source

PtFEM.sparin! — Method.

Function sparin! performs Cholesky factorisation on a symmetric skyline global matrix. The vector kv is updated.

Arguments to sparin!(kv, kdiag):

kv::Vector{Float64} : Skyline vector of global stiffness matrix

kdiag::Vector{Int64} : Diagonal fin_el vector

source

Index

PtFEM.Beam

PtFEM.FiniteElement

PtFEM.Line

PtFEM.Rod

PtFEM.StructuralElement

PtFEM.Triangle

Base.:*

PtFEM.fromSkyline

PtFEM.p41

PtFEM.sparin!