Programs

4 Static Equilibrium Programs

PtFEM.p41 — Function.

Method p41

One dimensional analysis of an axially loaded elastic Rod using 2-node Line elements.

Constructors

p41(data)

Arguments

* `data::Dict{Symbol, Any}`  : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Float64}                        : x-coordinate vector

Optional additional data dictionary keys

* penalty = 1e20               : Penalty used for fixed degrees of freedoms
* etype::Vector{Int64}         : Element material vector if np_types > 1
* eq_nodal_forces_and_moments  : Contribution of distributed loads to loaded_nodes

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement             : List of available structural element types
?Rod                           : Help on a Rod structural element
?FiniteElement                 : List finite element types
?Line                          : Help on Line finite element

source

Method p41

One dimensional analysis of an axially loaded elastic Rod using 2-node Line elements.

Constructors

p41(m, data) # Re-use factored global stiffness matrix

Arguments

* `m::jFEM`                  : Previously created jFEM model
* `data::Dict{Symbol, Any}`  : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Float64}                        : x-coordinate vector

Optional additional data dictionary keys

* penalty = 1e20               : Penalty used for fixed degrees of freedoms
* etype::Vector{Int64}         : Element material vector if np_types > 1
* eq_nodal_forces_and_moments  : Contribution of distributed loads to loaded_nodes

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement             : List of available structural element types
?Rod                           : Help on a Rod structural element
?FiniteElement                 : List finite element types
?Line                          : Help on Line finite element

source

PtFEM.p42 — Method.

Method p42

Analysis of elastic pin-jointed frames using 2-node rod elements in 2- or 3-dimensions.

Constructors

p42(data)

Arguments

* `data::Dict{Symbol, Any}`  : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of structural element
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::Vector{Float64}                            : x coordinate vector
* y_coords::Vector{Float64}                            : y coordinate vector
* g_num::Array{Int64,2}                                : Element node connections

Optional additional dictionary keys

* penalty::Float64             : Penalty for fixed freedoms
* etype::Vector{Int64}         : Element material vector
* z_coords::Vector{Float64}    : z coordinate vector (3D)
* eq_nodal_forces_and_moments  : Contribution of distributed loads to loaded_nodes

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement  : List structural element types
?Frame              : Help on a Rod structural fin_el
?FiniteElement      : List finite element types
?Line               : Help on Line finite element

source

PtFEM.p43 — Method.

p43

Analysis of elastic beams using 2-node Beam structural elements and Line finite elements. Elastic foundation is optional.

Constructors

p43(data)

Arguments

* `data::Dict{Symbol, Any}` : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::LinSpace{Float64}                          : x coordinate vector
* g_num::Array{Int64,2}                                : Element node connections
* fixed_freedoms::Array{Tuple{Vector{Int64}}           : Fixed freedoms

Optional additional dictionary keys

* etype::Vector{Int64}                                 : Element material vector
* penalty::Float64                                     : Penalty for fixed freedoms
* eq_nodal_forces_and_moments                          : Equivalent nodal loads

Return values

* (jfem, dis_dt, fm_dt)        : Tuple of jFem, dis_dt and fm_dt
                                 where:
                                    jfem::jFem    : Computational result type
                                    dis_dt        : Displacement data table
                                    fm_dt         : Forces and moments data table

Related help

?StructuralElement  : Help on structural elements
?Rod                : Help on a Rod structural fin_el
?FiniteElement      : Help on finite element types

source

PtFEM.p44 — Method.

p44

Analysis of elastic rigid-joonted frames using a 2-node Frame structural element and Line finite elements in 2 or 3 dimensions.

Constructors

p44(data)

Arguments

* `data::Dict{Symbol, Any}` : Dictionary containing all input data

Dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Float64}                        : x coordinate vector
* y_coords::FloatRange{Float64}                        : y coordinate vector
* g_num::Array{Int64,2}                                : Element node connections
* fixed_freedoms::Array{Tuple{Vector{Int64}}           : Fixed freedoms

Optional additional dictionary keys

* etype::Vector{Int64}                                 : Element material vector
* penalty::Float64                                     : Penalty for fixed freedoms
* z_coords::FloatRange{Float64}                        : z coordinate vector
* eq_nodal_forces_and_moments                          : Equivalent nodal loads

Related help

?StructuralElement  : Help on structural elements
?Beam               : Help on a Beam structural fin_el
?FiniteElement      : Help on finite element types

source

PtFEM.p45 — Method.

Method p45

Analysis of elasto-plastic beams or rigid-jointed frames using a 2-node Frame structural element in 1, 2 or 3 dimensions.

Constructors

p45(data)

Arguments

* `data::Dict{Symbol, Any}`  : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Type of  structural element
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Float64}                        : x-coordinate vector
* dload::FloatRange{Float64}                           : load steps

Optional additional data dictionary keys

* penalty = 1e20                 : Penalty used for fixed degrees of freedoms
* etype::Vector{Int64}           : Element material vector if np_types > 1
* y_coords::FloatRange{Float64}  : y-coordinate vector (2D)
* z_coords::FloatRange{Float64}  : x-coordinate vector (3D)
* limit = 250                    : Iteration limit
* tol = 0.0001                   : Tolerance for iteration convergence

Related help

?StructuralElement             : List of available structural element types
?Frame                         : Help on a Frame structural element
?FiniteElement                 : List finite element types
?Line                          : Help on Line finite element

source

PtFEM.p46 — Method.

Method p46

Stability (buckling) analysis of elastic beams using a 2-node Beam structural element and Line finite elements. Elastic foundation is optional.

Constructors

p46(data)

Arguments

* `data::Dict{Symbol, Any}` : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Type of  structural fin_el
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Float64}                        : x-coordinate vector

Optional additional data dictionary keys

* etype::Vector{Int64}         : Element material vector if np_types > 1
* limit = 250                  : Iteration limit
* tol = 0.0001                 : Tolerance for iteration convergence

Related help

?StructuralElement             : List of available structural element types
?Beam                          : Help on a Beam structural element
?FiniteElement                 : List finite element types
?Line                          : Help on Line finite element

source

PtFEM.p47 — Method.

Method p47

Analysis of plates (Plane structural element) using 4-node Quadrilateral finite elements. Homogeneous material with identical elements. Mesh numbered in x or y direction.

Constructors

p47(data)

Arguments

* `data::Dict{Symbol, Any}` : Dictionary containing all input data

Required data dictionary keys

* struc_el::StructuralElement                          : Structural element
* support::Array{Tuple{Int64,Array{Int64,1}},1}        : Fixed-displacements vector
* loaded_nodes::Array{Tuple{Int64,Array{Float64,1}},1} : Node load vector
* properties::Vector{Float64}                          : Material properties
* x_coords::FloatRange{Floalt64}                       : x-coordinate vector
* y_coords::FloatRange{Floalt64}                       : y-coordinate vector
* thickness:: Float64                                  : Thickness of plate

Optional additional data dictionary keys

* penalty = 1e20               : Penalty used for fixed degrees of freedoms
* etype::Vector{Int64}         : Element material vector if np_types > 1

Return values

* (fm_dt, sigma_dt)            : Tuple of jFem, dis_dt and fm_dt
                                  where:
                                    fm_dt         : Forces and moments data table
                                    sigma_dt      : Stresses data table

Related help

?StructuralElement             : List of available structural element types
?Plane                         : Help on a Plane structural element
?FiniteElement                 : List finite element types
?Quadrilateral                 : Help on Quadrilateral finite element

source

5 Elastic Solids Programs

p51(data::Dict{Symbol, Any})
p52(data::Dict{Symbol, Any})
p53(data::Dict{Symbol, Any})
p54(data::Dict{Symbol, Any})
p55(data::Dict{Symbol, Any})
p56(data::Dict{Symbol, Any})

6 Material Nonlinearity Programs

p61(data::Dict{Symbol, Any})
p62(data::Dict{Symbol, Any})

Structural Element Types

StructuralElement
Rod
Beam
Frame
Plane
Solid
GenericSolid

Finite Element Types

PtFEM.FiniteElement — Type.

FiniteElement

Abstract finite element type.

Type

abstract FiniteElement

Subtypes

* Line::FiniteElement          : 1D Line(nod, nodof)
* Triangle::FiniteElement      : 2D Triangle(nod, nodof)
* Quadrilateral::FiniteElement : 2D Quadrilateral(nod, nodof)
* Hexahedron::FiniteElement    : 3D Hexahedron(nod, nodof)
* Tetrahedron::FiniteElement   : 3D Tetrahedron(nod, nodof)

source

PtFEM.Line — Type.

Line (Interval)

1D type finite element

Constructor

Line(nod, nodof)
Line(nodof)

Arguments

* nod::Int64       : Number of nodes for finite element, defaults to 2
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Triangle — Type.

Triangle

2D type finite element

Constructor

Triangle(nod, nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (3, 6, 10, 15)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Quadrilateral — Type.

Quadrilateral

2D type finite element

Constructor

Quadrilateral(nod, nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (4, 8, 9)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Hexahedron — Type.

hexahedron

3D type finite element

Constructor

Hexahedron(nod, nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (8, 14, 20)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

PtFEM.Tetrahedron — Type.

Tetrahedron

3D type finite element

Constructor

Tetrahedron(nod, nodof)
Tetrahedron(nodof)

Arguments

* nod::Int64       : Number of nodes for finite element (defaults to 4)
* nodof::Int64     : Number of degrees of freedom per node

Related help

?FiniteElement      : Help on finite element types

source

Other Julia Types

FEM
jFEM

PtFEM - Main

PtFEM.beam_gm!(gm::Matrix{Float64}, ell::Float64)
PtFEM.beam_km!(km::Matrix{Float64}, ei::Float64, ell::Float64)
PtFEM.beam_mm!(mm::Matrix{Float64}, fs::Float64, ell::Float64)
PtFEM.beemat_nonaxi!(bee::Matrix{Float64}, radius::Float64, coord::Matrix{Float64}, deriv::Matrix{Float64}, fun::Vector{Float64}, iflag::Int64, lth::Int64)
PtFEM.beemat!(bee::Matrix{Float64},deriv::Matrix{Float64})
PtFEM.checon!(loads::Vector{Float64}, oldlds::Vector{Float64}, tol::Float64)
PtFEM.deemat!(dee::Array{Float64, 2}, e::Float64, v::Float64)
PtFEM.fkdiag!(ndof::Int64, neq::Int64, g::Vector{Int64}, kdiag::Vector{Int64})
PtFEM.fkdiag!(kdiag::Vector{Int64}, g::Vector{Int64})
PtFEM.fmplat!(d2x::Vector{Float64}, d2y::Vector{Float64}, d2xy::Vector{Float64}, points::Matrix{Float64}, aa::Float64, bb::Float64, i::Int64)
PtFEM.formm!(stress::Vector{Float64}, m1::Matrix{Float64}, m2::Matrix{Float64}, m3::Matrix{Float64})
PtFEM.formnf!(nodof::Int64, nn::Int64, nf::Matrix{Int64})
PtFEM.glob_to_loc!(loc::Vector{Float64}, glob::Vector{Float64}, gamma::Float64, coord::Matrix{Float64})
PtFEM.global_to_axial(glob::Vector{Float64}, coord::Matrix{Float64})
PtFEM.hinge!(coord::Matrix{Float64}, holdr::Matrix{Float64}, action::Vector{Float64}, react::Vector{Float64}, prop::Matrix{Float64}, iel, etype::Vector{Int64}, gamma::Vector{Float64})
PtFEM.invar!(stress::Vector{Float64}, sigm::Float64, dsbar::Float64, theta::Float64)
PtFEM.linmul_sky!(kv::Vector{Float64}, disps::Vector{Float64}, loads::Vector{Float64}, kdiag::Vector{Int64})
PtFEM.loc_to_glob!(loc::Vector{Float64}, glob::Vector{Float64}, gamma::Float64, coord::Matrix{Float64})
PtFEM.num_to_g!(num::Vector{Int64}, nf::Matrix{Int64}, g::Vector{Int64})
PtFEM.pin_jointed!(km::Matrix{Float64}, ea::Float64, coord::Matrix{Float64})
PtFEM.rigid_jointed!(km::Matrix{Float64}, prop::Matrix{Float64}, gamma::Vector{Float64}, etype::Vector{Int64}, iel::Int64, coord::Matrix{Float64})
PtFEM.rod_km!(km::Matrix{Float64}, ea::Float64, length::Float64)
PtFEM.format_nf_line!(l::AbstractString, t::Array{Tuple})
PtFEM.read_nf_file(f::AbstractString)
PtFEM.format_loads_line!(l::AbstractString, t::Array{Tuple})
PtFEM.read_loads_file(f::AbstractString)
PtFEM.rod_mm!(mm::Matrix{Float64}, length::Float64)
PtFEM.sample!(fin_el::Hexahedron, s::Matrix{Float64} , wt::Vector{Float64})
PtFEM.sample!(fin_el::Line, s::Matrix{Float64} , wt::Vector{Float64})
PtFEM.sample!(fin_el::Quadrilateral, s::Matrix{Float64} , wt::Vector{Float64})
PtFEM.sample!(fin_el::Tetrahedron, s::Matrix{Float64} , wt::Vector{Float64})
PtFEM.sample!(fin_el::Triangle, s::Matrix{Float64} , wt::Vector{Float64})
PtFEM.shape_der!(der::Matrix{Float64}, points::Matrix{Float64}, i::Int64)
PtFEM.shape_fun!(fun::Vector{Float64}, points::Matrix{Float64}, i::Int64)
PtFEM.stability(gsm::SparseMatrixCSC{Float64,Int64}, ggm::SparseMatrixCSC{Float64,Int64}, tol::Float64, limit::Int64)

PtFEM - Geom

PtFEM.geom_rect!(fin_el::Quadrilateral, iel::Int64, x_coords::Array{Float64, 1}, y_coords::Array{Float64, 1}, coord::Matrix{Float64}, num::Vector{Int64}, dir::Symbol)
PtFEM.geom_rect!(fin_el::Triangle, iel::Int64, x_coords::Array{Float64, 1}, y_coords::Array{Float64, 1}, coord::Matrix{Float64}, num::Vector{Int64}, dir::Symbol)
hexahedron_xz!(iel::Int64, x_coords::Vector{Float64}, y_coords::Vector{Float64}, z_coords::Vector{Float64}, coord::Matrix{Float64}, num::Vector{Int64})
PtFEM.mesh_size(fe::Line, nxe::Int64)
PtFEM.mesh_size(fe::Triangle, nxe::Int64, nye::Int64)
PtFEM.mesh_size(fe::Quadrilateral, nxe::Int64, nye::Int64)
PtFEM.mesh_size(fe::Hexahedron, nxe::Int64, nye::Int64, nze::Int64)

PtFEM - Plot methods

PtFEM.mesh — Method.

mesh

Plots displacements and directions

Function

mesh(data, g_coord, g_num, disp, ampl, pdir)

Arguments

* data::Dict                 : Input dictionary
* g_coord::Array{Float64, 2} : Coordinate array
* g_num::Array{Int, 2}       : Global node numbering array
* disp::DataTable            : Displacements DataTable
* ampl::Float64              : Amplification for derivatives
* pdir::AbstractString       : Nodal freedom matrix (updated)

source

PtFEM - VTK methods

PtFEM.vtk(data::Dict, fm_dt, sigma_dt, dir, fname)

PtFEM - Julia functions & operators

Base.LinAlg.cholfact — Function.

cholfact(A, [uplo::Symbol,] Val{false}) -> Cholesky

Compute the Cholesky factorization of a dense symmetric positive definite matrix A and return a Cholesky factorization. The matrix A can either be a Symmetric or Hermitian StridedMatrix or a perfectly symmetric or Hermitian StridedMatrix. In the latter case, the optional argument uplo may be :L for using the lower part or :U for the upper part of A. The default is to use :U. The triangular Cholesky factor can be obtained from the factorization F with: F[:L] and F[:U]. The following functions are available for Cholesky objects: size, \, inv, det. A PosDefException exception is thrown in case the matrix is not positive definite.

source

cholfact(A, [uplo::Symbol,] Val{true}; tol = 0.0) -> CholeskyPivoted

Compute the pivoted Cholesky factorization of a dense symmetric positive semi-definite matrix A and return a CholeskyPivoted factorization. The matrix A can either be a Symmetric or Hermitian StridedMatrix or a perfectly symmetric or Hermitian StridedMatrix. In the latter case, the optional argument uplo may be :L for using the lower part or :U for the upper part of A. The default is to use :U. The triangular Cholesky factor can be obtained from the factorization F with: F[:L] and F[:U]. The following functions are available for PivotedCholesky objects: size, \, inv, det, and rank. The argument tol determines the tolerance for determining the rank. For negative values, the tolerance is the machine precision.

source

cholfact(A; shift = 0.0, perm = Int[]) -> CHOLMOD.Factor

Compute the Cholesky factorization of a sparse positive definite matrix A. A must be a SparseMatrixCSC, Symmetric{SparseMatrixCSC}, or Hermitian{SparseMatrixCSC}. Note that even if A doesn't have the type tag, it must still be symmetric or Hermitian. A fill-reducing permutation is used. F = cholfact(A) is most frequently used to solve systems of equations with F\b, but also the methods diag, det, logdet are defined for F. You can also extract individual factors from F, using F[:L]. However, since pivoting is on by default, the factorization is internally represented as A == P'*L*L'*P with a permutation matrix P; using just L without accounting for P will give incorrect answers. To include the effects of permutation, it's typically preferable to extact "combined" factors like PtL = F[:PtL] (the equivalent of P'*L) and LtP = F[:UP] (the equivalent of L'*P).

Setting optional shift keyword argument computes the factorization of A+shift*I instead of A. If the perm argument is nonempty, it should be a permutation of 1:size(A,1) giving the ordering to use (instead of CHOLMOD's default AMD ordering).

Note

This method uses the CHOLMOD library from SuiteSparse, which only supports doubles or complex doubles. Input matrices not of those element types will be converted to SparseMatrixCSC{Float64} or SparseMatrixCSC{Complex128} as appropriate.

Many other functions from CHOLMOD are wrapped but not exported from the Base.SparseArrays.CHOLMOD module.

source

Base.:\ — Function.

\(x, y)

Left division operator: multiplication of y by the inverse of x on the left. Gives floating-point results for integer arguments.

source

\(A, B)

Matrix division using a polyalgorithm. For input matrices A and B, the result X is such that A*X == B when A is square. The solver that is used depends upon the structure of A. If A is upper or lower triangular (or diagonal), no factorization of A is required and the system is solved with either forward or backward substitution. For non-triangular square matrices, an LU factorization is used.

For rectangular A the result is the minimum-norm least squares solution computed by a pivoted QR factorization of A and a rank estimate of A based on the R factor.

When A is sparse, a similar polyalgorithm is used. For indefinite matrices, the LDLt factorization does not use pivoting during the numerical factorization and therefore the procedure can fail even for invertible matrices.

source

PtFEM - Parallel processing

Base.pmap — Function.

pmap([::AbstractWorkerPool], f, c...; distributed=true, batch_size=1, on_error=nothing, retry_n=0, retry_max_delay=DEFAULT_RETRY_MAX_DELAY, retry_on=DEFAULT_RETRY_ON) -> collection

Transform collection c by applying f to each element using available workers and tasks.

For multiple collection arguments, apply f elementwise.

Note that f must be made available to all worker processes; see Code Availability and Loading Packages for details.

If a worker pool is not specified, all available workers, i.e., the default worker pool is used.

By default, pmap distributes the computation over all specified workers. To use only the local process and distribute over tasks, specify distributed=false. This is equivalent to asyncmap.

pmap can also use a mix of processes and tasks via the batch_size argument. For batch sizes greater than 1, the collection is split into multiple batches, which are distributed across workers. Each such batch is processed in parallel via tasks in each worker. The specified batch_size is an upper limit, the actual size of batches may be smaller and is calculated depending on the number of workers available and length of the collection.

Any error stops pmap from processing the remainder of the collection. To override this behavior you can specify an error handling function via argument on_error which takes in a single argument, i.e., the exception. The function can stop the processing by rethrowing the error, or, to continue, return any value which is then returned inline with the results to the caller.

Failed computation can also be retried via retry_on, retry_n, retry_max_delay, which are passed through to retry as arguments retry_on, n and max_delay respectively. If batching is specified, and an entire batch fails, all items in the batch are retried.

The following are equivalent:

pmap(f, c; distributed=false) and asyncmap(f,c)

pmap(f, c; retry_n=1) and asyncmap(retry(remote(f)),c)

pmap(f, c; retry_n=1, on_error=e->e) and asyncmap(x->try retry(remote(f))(x) catch e; e end, c)

source

PtFEM - No longer used

PtFEM.sparin!(kv::Vector{Float64}, kdiag::Vector{Int64})
PtFEM.spabac!(kv::Vector{Float64}, loads::Vector{Float64}, kdiag::Vector{Int64})
fromSkyline(skyline::Vector{Float64}, kdiag::Vector{Int64})
PtFEM.skyline2sparse(skyline::Vector{Float64}, kdiag::Vector{Int64})
PtFEM.stability!(kv::Vector{Float64}, gv::Vector{Float64}, kdiag::Vector{Int64}, tol::Float64, limit::Int64, iters::Int64, evec::Vector{Float64}, ival::Float64)

Index

PtFEM.Beam

PtFEM.FiniteElement

PtFEM.Hexahedron

PtFEM.Line

PtFEM.Quadrilateral

PtFEM.Rod

PtFEM.StructuralElement

PtFEM.Tetrahedron

PtFEM.Triangle

Base.:\

Base.LinAlg.cholfact

Base.pmap

PtFEM.formnf!

PtFEM.mesh

PtFEM.mesh_size

PtFEM.mesh_size

PtFEM.mesh_size

PtFEM.mesh_size

PtFEM.p41

PtFEM.p42

PtFEM.p43

PtFEM.p44

PtFEM.p45

PtFEM.p46

PtFEM.p47

PtFEM.spabac!

PtFEM.sparin!